feat: Implement initial login and registration pages using Mantine UI.
